Add redirect shortlink and admin api points

+++ backend/api/serializers.py
@@ -2,16 +2,14 @@ from django.contrib.auth import get_user_model
 from django.contrib.auth.hashers import make_password
 from django.db import transaction
 from rest_framework import serializers
-from rest_framework.exceptions import NotFound, AuthenticationFailed
-from rest_framework.serializers import ValidationError
 from drf_extra_fields.fields import Base64ImageField
-from rest_framework.response import Response
 
-from foodgram_backend.settings import REST_FRAMEWORK
 from recipe.models import (Recipe, Tag, Ingredient, RecipeIngredient,
                            RecipeFavorite, RecipeShoppingCart)
 from user.models import Follow
 from .mixins import PasswordChangeMixin, PasswordMixin
+from .validation import (validate_tags_and_ingredients,
+                         validate_ingredient_data, validate_recipes_limit)
 
 User = get_user_model()
 
@@ -44,15 +42,37 @@ class UserCreateSerializer(PasswordMixin):
         )
         extra_kwargs = {'password': {'write_only': True}}
 
-    def create(self, validated_data):
-        encrypt_password = make_password(validated_data['password'])
-        validated_data['password'] = encrypt_password
-        return super(UserCreateSerializer, self).create(validated_data)
 
+class UserFollowSerializer(serializers.ModelSerializer):
+    recipes = serializers.SerializerMethodField()
+    is_subscribed = serializers.SerializerMethodField()
+    recipes_count = serializers.IntegerField(read_only=True)
 
-class PasswordChangeSerializer(PasswordChangeMixin):
-    current_password = serializers.CharField(required=True)
-    new_password = serializers.CharField(required=True)
+    class Meta:
+        model = User
+        fields = (
+            'email', 'id', 'username', 'first_name',
+            'last_name', 'is_subscribed', 'recipes', 'recipes_count', 'avatar'
+        )
+
+    def get_recipes(self, obj):
+        request = self.context.get('request')
+        recipes = obj.recipes.all()
+        recipes_limit = validate_recipes_limit(request)
+        if recipes_limit is not None:
+            recipes = recipes[:recipes_limit]
+        serializer = RecipeShortSerializer(
+            recipes, many=True, read_only=True
+        )
+        return serializer.data
+
+    def get_is_subscribed(self, obj):
+        request = self.context.get('request')
+        follower = request.user
+        if request and follower.is_authenticated:
+            follow = Follow.objects.filter(follower=follower, following=obj)
+            return follow.exists()
+        return False
 
 
 class UserAvatarSerializer(serializers.ModelSerializer):
@@ -61,6 +81,11 @@ class UserAvatarSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
         fields = ('avatar',)
+
+
+class PasswordChangeSerializer(PasswordChangeMixin):
+    current_password = serializers.CharField(required=True)
+    new_password = serializers.CharField(required=True)
 
 
 class TagSerializer(serializers.ModelSerializer):
@@ -118,24 +143,8 @@ class RecipeSerializer(serializers.ModelSerializer):
 
     @staticmethod
     def create_ingredient(instance, item_data):
-        if not item_data:
-            raise serializers.ValidationError(
-                'Поле c ингредиентами не может быть пустым'
-            )
-        try:
-            ing_id = item_data.get('ingredient', {}).get('id')
-            amount = item_data.get('amount')
-            if not ing_id or not amount:
-                raise serializers.ValidationError(
-                    'Поле c ингредиентами не заполнено'
-                )
-            if amount < 1:
-                raise serializers.ValidationError(
-                    'Количество ингредиента должно быть больше нуля'
-                )
-            exist_ingredient = Ingredient.objects.get(id=ing_id)
-        except Ingredient.DoesNotExist:
-            raise NotFound(f'Ингредиент не найден')
+        ing_id, amount = validate_ingredient_data(item_data)
+        exist_ingredient = Ingredient.objects.get(id=ing_id)
         ingredient, created = RecipeIngredient.objects.get_or_create(
             recipe=instance,
             ingredient=exist_ingredient,
@@ -146,44 +155,50 @@ class RecipeSerializer(serializers.ModelSerializer):
             ingredient.save()
         return ingredient
 
-    def create(self, validated_data):
-        if self.context.get('request').user.is_anonymous:
-            raise AuthenticationFailed('Пользователь не авторизован')
-        with transaction.atomic():
-            ingredients_data, tags_data = self.pop_items(validated_data)
+    def set_ingredients_tags(self, validated_data, instance=None):
+        request = self.context.get('request')
+        ingredients_data, tags_data = self.pop_items(validated_data)
+        validate_tags_and_ingredients(request, ingredients_data, tags_data)
+        if request.method == 'POST':
             recipe = Recipe.objects.create(**validated_data)
-            recipe.tags.set(tags_data)
-            for ingredient_data in ingredients_data:
-                self.create_ingredient(recipe, ingredient_data)
-        return recipe
+            instance = recipe
+        if tags_data:
+            instance.tags.set(tags_data)
+        if request.method == 'PATCH':
+            if ingredients_data:
+                instance.recipe_ingredients.all().delete()
+            for attr, value in validated_data.items():
+                setattr(instance, attr, value)
+        for ingredient_data in ingredients_data:
+            self.create_ingredient(instance, ingredient_data)
+        instance.save()
+        return instance
+
+    def create(self, validated_data):
+        with transaction.atomic():
+            return self.set_ingredients_tags(validated_data)
 
     def update(self, instance, validated_data):
         with transaction.atomic():
-            ingredients_data, tags_data = self.pop_items(validated_data)
-            instance.tags.set(tags_data)
-            instance.recipe_ingredients.all().delete()
-            for attr, value in validated_data.items():
-                setattr(instance, attr, value)
-            for ingredient_data in ingredients_data:
-                self.create_ingredient(instance, ingredient_data)
-            instance.save()
-        return instance
+            return self.set_ingredients_tags(validated_data, instance)
+
+    @staticmethod
+    def exist_in_queryset(user, model, obj):
+        if user.is_authenticated:
+            return model.objects.filter(user=user, recipe=obj).exists()
+        return False
 
     def get_is_favorited(self, obj):
         request = self.context.get('request')
-        user = request.user
-        if request and user.is_authenticated:
-            favorite = RecipeFavorite.objects.filter(user=user, recipe=obj)
-            return favorite.exists()
-        return False
+        if not request:
+            return False
+        return self.exist_in_queryset(request.user, RecipeFavorite, obj)
 
     def get_is_in_shopping_cart(self, obj):
         request = self.context.get('request')
-        user = request.user
-        if request and user.is_authenticated:
-            shopping = RecipeShoppingCart.objects.filter(user=user, recipe=obj)
-            return shopping.exists()
-        return False
+        if not request:
+            return False
+        return self.exist_in_queryset(request.user, RecipeShoppingCart, obj)
 
     def to_representation(self, instance):
         representation = super().to_representation(instance)
@@ -192,42 +207,7 @@ class RecipeSerializer(serializers.ModelSerializer):
         return representation
 
 
-class RecipeFavoriteSerializer(serializers.ModelSerializer):
+class RecipeShortSerializer(serializers.ModelSerializer):
     class Meta:
         model = Recipe
         fields = ('id', 'name', 'image', 'cooking_time')
-
-
-class UserFollowSerializer(serializers.ModelSerializer):
-    recipes = serializers.SerializerMethodField()
-    is_subscribed = serializers.SerializerMethodField()
-    recipes_count = serializers.IntegerField(read_only=True)
-
-    class Meta:
-        model = User
-        fields = (
-            'email', 'id', 'username', 'first_name',
-            'last_name', 'is_subscribed', 'recipes', 'recipes_count', 'avatar'
-        )
-
-    def get_recipes(self, obj):
-        request = self.context.get('request')
-        recipes = obj.recipes.all()
-        if 'recipes_limit' in request.query_params:
-            try:
-                recipes_limit = int(request.query_params['recipes_limit'])
-            except (TypeError, ValueError):
-                raise ValidationError('recipes_limit должен быть целым числом')
-            recipes = obj.recipes.all()[:recipes_limit]
-        serializer = RecipeFavoriteSerializer(
-            recipes, many=True, read_only=True
-        )
-        return serializer.data
-
-    def get_is_subscribed(self, obj):
-        request = self.context.get('request')
-        follower = request.user
-        if request and follower.is_authenticated:
-            follow = Follow.objects.filter(follower=follower, following=obj)
-            return follow.exists()
-        return False
